Isolate tekken save_format into TekkenBackend instead of TokenizersBackend
