+++ src/transformers/integrations/mistral/__init__.py
@@ -22,6 +22,7 @@ from ...utils import _LazyModule
 _import_structure = {
     "tokenizer": [
         "MistralConverter",
+        "TekkenBackend",
         "convert_tekken_tokenizer",
         "resolve_mistral_format",
     ],
@@ -30,6 +31,7 @@ _import_structure = {
 if TYPE_CHECKING:
     from .tokenizer import (
         MistralConverter,
+        TekkenBackend,
         convert_tekken_tokenizer,
         resolve_mistral_format,
     )
